+++ lib/ui/RootPanel.dart
@@ -27,6 +27,7 @@ import 'package:illinois/service/DeviceCalendar.dart';
 import 'package:illinois/service/ExploreService.dart';
 import 'package:illinois/service/FlexUI.dart';
 import 'package:illinois/service/FirebaseMessaging.dart';
+import 'package:illinois/service/Groups.dart';
 import 'package:illinois/service/Polls.dart';
 import 'package:illinois/service/Service.dart';
 import 'package:illinois/service/Sports.dart';
@@ -37,6 +38,7 @@ import 'package:illinois/service/NotificationService.dart';
 import 'package:illinois/ui/SavedPanel.dart';
 import 'package:illinois/ui/athletics/AthleticsGameDetailPanel.dart';
 import 'package:illinois/ui/explore/ExplorePanel.dart';
+import 'package:illinois/ui/groups/GroupDetailPanel.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/BrowsePanel.dart';
 import 'package:illinois/ui/athletics/AthleticsHomePanel.dart';
@@ -101,6 +103,7 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
       FirebaseMessaging.notifyPopupMessage,
       FirebaseMessaging.notifyEventDetail,
       FirebaseMessaging.notifyAthleticsGameStarted,
+      FirebaseMessaging.notifyGroupsNotification,
       ExploreService.notifyEventDetail,
       Sports.notifyGameDetail,
       Localization.notifyStringsUpdated,
@@ -193,6 +196,9 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
     }
     else if (name == Polls.notifyPresentResult) {
       _presentPollResult(param);
+    }
+    else if (name == FirebaseMessaging.notifyGroupsNotification) {
+      _onFirebaseGroupsNotification(param);
     }
   }
 
@@ -550,6 +556,21 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
     }
     else {
       return null;
+    }
+  }
+
+  void _onFirebaseGroupsNotification(param) {
+    if(param is Map<String, dynamic>){
+      String groupId = param["entity_id"];
+      if(groupId != null) {
+        Groups().loadGroup(groupId).then((value){
+          if(value != null){
+            Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupDetailPanel(group: value,)));
+          }
+        }).catchError((err){
+          AppAlert.showDialogResult(context, Localization().getStringEx("panel.group_detail.label.error_message", "Failed to load group data."));
+        });
+      }
     }
   }